Add merge and pull sync mechanisms
This appears to be the cleanest way to merge upstream branches into the
destination, removing all extra files and overwriting all content.

The pull is simple.

Needs error handling not to explode.

Also adding a --dry-run option to the script that should be passed to
functions to:

a) not push anything to the destination SCM,
b) not submit any builds, and
c) not upload anything to the destination cache (verifying downloads is
fine, I think)

Also fixing a syntax error, heh.

Signed-off-by: Petr Šabata <contyk@redhat.com>