Implement the service mode
Untested but it should work.  Hopefully.

Messaging API requires callbacks with a single argument, so we need to
pass dry_run via a property.  It's ugly but it's what it is.

Since we need this, we could make all the other functions use the same
thing, too.  And make APIs for setting and getting it.  Later.

Signed-off-by: Petr Šabata <contyk@redhat.com>