Allow for synchronizing selected configured components
Adding the -s/--select option to distrobaker which accepts
space-separated list of configured components to only sync a subset of
the configured components.

Useful for keeping a large configuration file and only re-syncing a few
manually, or generic testing with the same configuration.

Currently only work with the oneshot mode.  Service mode should just use
a different configuration file.

Signed-off-by: Petr Šabata <contyk@redhat.com>