New features, config format changes
All these changes are intertwined, landing in one go.

Configuration:

The config file now supports a switch for the mode of operation, strict.
In the strict mode DistroBaker functions like it did before, only
processing configured components.  With strict: false, DistroBaker
accepts inputs even for components that are not explicitly configured,
applying defaults (see below).

Additionaly, the control block now supports an explicit list of excluded
components.  Listed components will be skipped, regardless of the mode
and even if they are explicitly configured.

The new defaults block configures the default SCM and cache values for
all components.  This implies component definitions are now optional.
They may still override the defaults if necessary.

Also, components may now define their lookaside cache names for both the
source and the destination.

Features:

Besides plenty of configuration options now being optional, DistroBaker
now queries for specific NVRs rather than blindly syncing the configured
branches.

When a trigger message is received, we look up the actual SCMURL used to
build that artifact and sync from there.  If running in the oneshot mode
with no --select list, we look up all builds tagged in the trigger tag
and process those.  If a --select list is specified, we limit ourselves
to the provided list.  Excluded components are skipped.

If source branch is specified, DistroBaker pulls only that particular
branch to locate the SCMURL commit.  If it's not, it fetches all remote
branches.  This is safer but slower.

Signed-off-by: Petr Šabata <contyk@redhat.com>