Submit builds using the configured destination component
Previously we submitted builds using the component name, which is not
correct if the component definition overrides the destination SCM.

As of this change, we now extract the build component name from the
destination SCM definition.

As a consequence, the get_scmurl() function now attempts to extract
namespaces and component names from the provided string.

Signed-off-by: Petr Šabata <contyk@redhat.com>